+++ python/shipDet_conf.py
@@ -701,8 +701,9 @@ def configure(run, ship_geo):
             ship_geo.strawtubes.DeltazFrame = 10.0 * u.cm
             ship_geo.strawtubes.FrameLateralWidth = 1.0 * u.cm
             ship_geo.strawtubes.FrameMaterial = "aluminium"
+        ship_geo.strawtubes.medium = "vacuums" if ship_geo.DecayVolumeMedium == "vaccums" else "air"
 
-        Strawtubes = ROOT.strawtubes("Strawtubes", ROOT.kTRUE)
+        Strawtubes = ROOT.strawtubes(ship_geo.strawtubes.medium)
         Strawtubes.SetZpositions(
             ship_geo.vetoStation.z,
             ship_geo.TrackStation1.z,